refactor: move texture slot manager to quad renderer
